extract Settings page from renderer
